Moved QRCodeEncodeDecodeTest to qrcode module
