+++ qr-code/src/test/java/org/odk/collect/qrcode/QRCodeEncodeDecodeTest.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.configure.qr
+package org.odk.collect.qrcode
 
 import android.graphics.Bitmap
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -6,8 +6,6 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.odk.collect.qrcode.QRCodeEncoderImpl
-import org.odk.collect.qrcode.QRCodeDecoderImpl
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
